requirements.txt
# Requirements
gym
numpy
pyyaml